Reversed local procedures (embedding model is too big, RAG is out of scope)
