+++ interpreter/core/core.py
@@ -13,7 +13,6 @@ from interpreter.utils import display_markdown_message
 
 from ..cli.cli import cli
 from ..llm.setup_llm import setup_llm
-from ..rag.get_relevant_procedures_string import get_relevant_procedures_string
 from ..terminal_interface.terminal_interface import terminal_interface
 from ..terminal_interface.validate_llm_settings import validate_llm_settings
 from ..utils.check_for_update import check_for_update
@@ -59,14 +58,6 @@ class Interpreter:
         self.max_budget = None
         self._llm = None
         self.gguf_quality = None
-
-        # Procedures / RAG
-        self.procedures = None
-        self._procedures_db = {}
-        self.download_open_procedures = True
-        self.embed_function = embed_function
-        # Number of procedures to add to the system message
-        self.num_procedures = 2
 
         # Load config defaults
         self.extend_config(self.config_file)
@@ -161,11 +152,8 @@ class Interpreter:
             code_interpreter.terminate()
         self._code_interpreters = {}
 
-        # Reset the two functions below, in case the user set them
+        # Reset the function below, in case the user set it
         self.generate_system_message = lambda: generate_system_message(self)
-        self.get_relevant_procedures_string = lambda: get_relevant_procedures_string(
-            self
-        )
 
         self.__init__()
 
@@ -173,6 +161,3 @@ class Interpreter:
     # I wish we could just dynamically expose all of our functions to devs...
     def generate_system_message(self):
         return generate_system_message(self)
-
-    def get_relevant_procedures_string(self):
-        return get_relevant_procedures_string(self)
