+++ tests/test_interpreter.py
@@ -1,14 +1,15 @@
 import os
+import re
 import time
 from random import randint
-import re
+
 import pytest
 
 import interpreter
 from interpreter.utils.count_tokens import count_messages_tokens, count_tokens
 
 interpreter.auto_run = True
-interpreter.model = "gpt-4"
+interpreter.model = "gpt-3.5-turbo"
 interpreter.temperature = 0
 
 
@@ -111,7 +112,7 @@ def test_hello_world():
     ]
 
 
-@pytest.mark.skip(reason="Math is hard")
+# @pytest.mark.skip(reason="Math is hard")
 def test_math():
     # we'll generate random integers between this min and max in our math tests
     min_number = randint(1, 99)
@@ -139,22 +140,53 @@ def test_delayed_exec():
     )
 
 
-@pytest.mark.skip(
-    reason="This works fine when I run it but fails frequently in Github Actions... will look into it after the hackathon"
-)
+# @pytest.mark.skip(
+#     reason="This works fine when I run it but fails frequently in Github Actions... will look into it after the hackathon"
+# )
 def test_nested_loops_and_multiple_newlines():
     interpreter.chat(
         """Can you write a nested for loop in python and shell and run them? Don't forget to properly format your shell script and use semicolons where necessary. Also put 1-3 newlines between each line in the code. Only generate and execute the code. No explanations. Thanks!"""
     )
 
+
 @pytest.mark.skip(
-    reason="Skipping until can verify it runs on the Github build server"
+    reason="This is too many tests (one for each parameter below) and I think that num_fibs runs len() on an int? We may want to delete this."
 )
-@pytest.mark.parametrize("expected_integers", [
-    0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144,
-    233, 377, 610, 987, 1597, 2584, 4181, 6765, 10946, 17711,
-    28657, 46368, 75025, 121393, 196418, 317811, 514229
-])
+@pytest.mark.parametrize(
+    "expected_integers",
+    [
+        0,
+        1,
+        1,
+        2,
+        3,
+        5,
+        8,
+        13,
+        21,
+        34,
+        55,
+        89,
+        144,
+        233,
+        377,
+        610,
+        987,
+        1597,
+        2584,
+        4181,
+        6765,
+        10946,
+        17711,
+        28657,
+        46368,
+        75025,
+        121393,
+        196418,
+        317811,
+        514229,
+    ],
+)
 def test_powershell_gen_exec(expected_integers):
     TASK_RESULT_PREFIX = "TaskResult:"
     num_fibs = str(len(expected_integers))  # Convert to string
@@ -166,22 +198,31 @@ def test_powershell_gen_exec(expected_integers):
         Make sure to always execute the Powershell code and provide the output in the following format: {TASK_RESULT_PREFIX} x0, x1, x2,...x(n)"""
     )
 
-    is_valid = lambda d: d.get('role') == 'assistant' and d.get('message', '').startswith(TASK_RESULT_PREFIX)
+    is_valid = lambda d: d.get("role") == "assistant" and d.get(
+        "message", ""
+    ).startswith(TASK_RESULT_PREFIX)
     valid_result = next(filter(is_valid, result), None)
-    
+
     if valid_result is not None:
-        message = valid_result.get('message', '')
-        pattern = r'{}\s*([\d\s,]+)'.format(re.escape(TASK_RESULT_PREFIX))  # Use re.escape() to handle any special characters
+        message = valid_result.get("message", "")
+        pattern = r"{}\s*([\d\s,]+)".format(
+            re.escape(TASK_RESULT_PREFIX)
+        )  # Use re.escape() to handle any special characters
         match = re.search(pattern, message)
         if match:
             integer_series = match.group(1)
-            extracted_integers = [int(num.strip()) for num in integer_series.split(',')]
-            
-            assert extracted_integers == expected_integers, "Extracted integers do not match expected integers."
-            
+            extracted_integers = [int(num.strip()) for num in integer_series.split(",")]
+
+            assert (
+                extracted_integers == expected_integers
+            ), "Extracted integers do not match expected integers."
+
             print(f"Extracted message for {num_fibs} Fibonacci numbers:", message)
     else:
-        print(f"No valid message found in the list of results for {num_fibs} Fibonacci numbers.")
+        print(
+            f"No valid message found in the list of results for {num_fibs} Fibonacci numbers."
+        )
+
 
 def test_markdown():
     interpreter.chat(
